+++ seqr/utils/search/elasticsearch/es_search.py
@@ -14,7 +14,7 @@ from seqr.utils.search.constants import XPOS_SORT_KEY, COMPOUND_HET, RECESSIVE, 
 from seqr.utils.search.elasticsearch.constants import X_LINKED_RECESSIVE, \
     HAS_ALT_FIELD_KEYS, GENOTYPES_FIELD_KEY, POPULATION_RESPONSE_FIELD_CONFIGS, POPULATIONS, \
     SORTED_TRANSCRIPTS_FIELD_KEY, CORE_FIELDS_CONFIG, NESTED_FIELDS, PREDICTION_FIELDS_CONFIG, INHERITANCE_FILTERS, \
-    QUERY_FIELD_NAMES, REF_REF, ANY_AFFECTED, GENOTYPE_QUERY_MAP, CLINVAR_SIGNFICANCE_MAP, HGMD_CLASS_MAP, \
+    QUERY_FIELD_NAMES, REF_REF, ANY_AFFECTED, GENOTYPE_QUERY_MAP, HGMD_CLASS_MAP, \
     SORT_FIELDS, MAX_VARIANTS, MAX_COMPOUND_HET_GENES, MAX_INDEX_NAME_LENGTH, QUALITY_QUERY_FIELDS, \
     GRCH38_LOCUS_FIELD, MAX_SEARCH_CLAUSES, SV_SAMPLE_OVERRIDE_FIELD_CONFIGS, \
     PREDICTION_FIELD_LOOKUP, SPLICE_AI_FIELD, CLINVAR_KEY, HGMD_KEY, CLINVAR_PATH_SIGNIFICANCES, \
@@ -141,8 +141,8 @@ class EsSearch(object):
                     self.samples_by_family_index[alias_index] = {}
                 self.samples_by_family_index[alias_index].update(alias_samples)
 
-    def _filter_families_for_inheritance(self, inheritance_filter):
-        for index, family_samples in list(self.samples_by_family_index.items()):
+    def _set_family_affected_status(self, inheritance_filter):
+        for family_samples in list(self.samples_by_family_index.values()):
             for family_guid, samples_by_id in family_samples.items():
                 individual_affected_status = _get_family_affected_status(samples_by_id, inheritance_filter)
                 if family_guid not in self._family_individual_affected_status:
@@ -213,17 +213,17 @@ class EsSearch(object):
 
         inheritance_filter = inheritance_filter or {}
 
-        if inheritance_mode or inheritance_filter:
-            self._filter_families_for_inheritance(inheritance_filter)
+        quality_affected_only = quality_filter and quality_filter.get('affected_only')
+        if inheritance_mode or inheritance_filter or quality_affected_only:
+            self._set_family_affected_status(inheritance_filter)
 
-        quality_filters_by_family = self._get_quality_filters_by_family(quality_filter)
+        quality_filters_by_family = self._get_quality_filters_by_family(quality_filter, quality_affected_only)
 
         comp_het_dataset_type = dataset_type
         if dataset_type != secondary_dataset_type:
             comp_het_dataset_type = None
         if dataset_type == ALL_DATA_TYPES:
             dataset_type = None
-        if comp_het_dataset_type == ALL_DATA_TYPES:
             comp_het_dataset_type = None
 
         has_comp_het_search = inheritance_mode in {RECESSIVE, COMPOUND_HET} and not self.previous_search_results.get('grouped_results')
@@ -1028,7 +1028,7 @@ class EsSearch(object):
             all_gene_consequences = []
             if variant.get('svType'):
                 all_gene_consequences.append(variant['svType'])
-            if variant.get(CLINVAR_KEY, {}).get('clinicalSignificance') in self._consequence_overrides.get(CLINVAR_KEY, []):
+            if _is_matched_clinvar_significance(variant.get(CLINVAR_KEY, {}).get('clinicalSignificance'), self._consequence_overrides.get(CLINVAR_KEY)):
                 all_gene_consequences.append(CLINVAR_KEY)
             if variant.get(HGMD_KEY, {}).get('class') in self._consequence_overrides.get(HGMD_KEY, []):
                 all_gene_consequences.append(HGMD_KEY)
@@ -1234,7 +1234,7 @@ class EsSearch(object):
                 long_running.append({'task': task, 'parent_task_id': parent_id})
         return long_running
 
-    def _get_quality_filters_by_family(self, quality_filter):
+    def _get_quality_filters_by_family(self, quality_filter, affected_only):
         quality_field_configs = {
             'min_{}'.format(field): {'field': field, 'step': step} for field, step in QUALITY_QUERY_FIELDS.items()
         }
@@ -1250,7 +1250,15 @@ class EsSearch(object):
             for index in self._indices:
                 family_samples_by_id = self.samples_by_family_index[index]
                 for family_guid, samples_by_id in family_samples_by_id.items():
-                    family_sample_ids[family_guid].update(samples_by_id.keys())
+                    if affected_only:
+                        family_affected_status = self._family_individual_affected_status[family_guid]
+                        sample_ids = {
+                            sample_id for sample_id, sample in samples_by_id.items()
+                            if family_affected_status[sample.individual.guid] == Individual.AFFECTED_STATUS_AFFECTED
+                        }
+                    else:
+                        sample_ids = samples_by_id.keys()
+                    family_sample_ids[family_guid].update(sample_ids)
 
             path_filter = self._get_clinvar_pathogenic_override_filter()
             for family_guid, sample_ids in sorted(family_sample_ids.items()):
@@ -1308,7 +1316,7 @@ def _family_genotype_inheritance_filter(inheritance_mode, inheritance_filter, sa
     individual_genotype_filter = inheritance_filter.get('genotype') or {}
 
     if inheritance_mode == X_LINKED_RECESSIVE:
-        samples_q = Q('match', contig='X')
+        samples_q = Q('range', xpos={'gte': get_xpos('X', 1), 'lte': get_xpos('Y', 1)})
         for individual in individuals:
             if individual_affected_status[individual.guid] == Individual.AFFECTED_STATUS_UNAFFECTED \
                     and individual.sex == Individual.SEX_MALE:
@@ -1404,29 +1412,47 @@ def _parse_pathogenicity_filter(pathogenicity):
     clinvar_filters = pathogenicity.get(CLINVAR_KEY, [])
     hgmd_filters = pathogenicity.get(HGMD_KEY, [])
 
-    clinvar_clinical_significance_terms = set()
-    if clinvar_filters:
-        for clinvar_filter in clinvar_filters:
-            clinvar_clinical_significance_terms.update(CLINVAR_SIGNFICANCE_MAP.get(clinvar_filter, []))
-
     hgmd_class = set()
     if hgmd_filters:
         for hgmd_filter in hgmd_filters:
             hgmd_class.update(HGMD_CLASS_MAP.get(hgmd_filter, []))
 
-    return sorted(clinvar_clinical_significance_terms), sorted(hgmd_class)
+    return sorted(clinvar_filters), sorted(hgmd_class)
 
 
-def _pathogenicity_filter(clinvar_terms, hgmd_classes=None):
+VUS_FILTER = 'vus_or_conflicting'
+VUS_REGEX = 'Conflicting_interpretations_of_pathogenicity.*|~((.*[Bb]enign.*)|(.*[Pp]athogenic.*))'
+
+
+def _pathogenicity_filter(clinvar_filters, hgmd_classes=None):
     pathogenicity_filter = None
-    if clinvar_terms:
-        pathogenicity_filter = Q('terms', clinvar_clinical_significance=clinvar_terms)
+
+    path_regex = '|'.join([
+        VUS_REGEX if clinvar_filter == VUS_FILTER else f'.*{clinvar_filter.capitalize()}.*'
+        for clinvar_filter in clinvar_filters or []
+    ])
+    if path_regex:
+        pathogenicity_filter = Q('regexp', clinvar_clinical_significance=path_regex)
 
     if hgmd_classes:
         hgmd_q = Q('terms', hgmd_class=hgmd_classes)
         pathogenicity_filter = pathogenicity_filter | hgmd_q if pathogenicity_filter else hgmd_q
 
     return pathogenicity_filter
+
+
+def _is_matched_clinvar_significance(clinical_significance, clinvar_filters):
+    if not (clinical_significance and clinvar_filters):
+        return False
+
+    if VUS_FILTER in clinvar_filters:
+        exclude = [
+            path.capitalize() for path in ['pathogenic',  'likely_pathogenic', 'likely_benign',  'benign']
+            if path not in clinvar_filters
+        ]
+        return all(substring not in clinical_significance for substring in exclude)
+
+    return any(clinvar_filter.capitalize() in clinical_significance for clinvar_filter in clinvar_filters)
 
 
 def _annotations_filter(vep_consequences):