+++ seqr/utils/search/hail_search_utils.py
@@ -1,25 +1,32 @@
 from collections import defaultdict
-from django.db.models import F
+from django.db.models import F, Min
 
 import requests
 from reference_data.models import Omim, GeneConstraint, GENOME_VERSION_LOOKUP
-from seqr.models import Individual, Sample, PhenotypePrioritization
-from seqr.utils.search.constants import RECESSIVE, COMPOUND_HET, MAX_NO_LOCATION_COMP_HET_FAMILIES, PRIORITIZED_GENE_SORT
-from seqr.views.utils.orm_to_json_utils import get_json_for_queryset
+from seqr.models import Sample, PhenotypePrioritization
+from seqr.utils.search.constants import PRIORITIZED_GENE_SORT
+from seqr.utils.xpos_utils import MIN_POS, MAX_POS
+from settings import HAIL_BACKEND_SERVICE_HOSTNAME, HAIL_BACKEND_SERVICE_PORT
+
+HAIL_BACKEND_SERVICE_HOSTNAME = 'http://hail-search:5000/'
+
+def _hail_backend_url(path):
+    return f'{HAIL_BACKEND_SERVICE_HOSTNAME}:{HAIL_BACKEND_SERVICE_PORT}/{path}'
+
+
+def ping_hail_backend():
+    requests.get(_hail_backend_url('status'), timeout=5).raise_for_status()
 
 
 def get_hail_variants(samples, search, user, previous_search_results, genome_version, sort=None, page=1, num_results=100,
                       gene_agg=False, **kwargs):
 
-    sample_data = _get_sample_data(samples, search.get('inheritance_filter'))
-
     end_offset = num_results * page
     search_body = {
         'requester_email': user.email,
-        'sample_data': sample_data,
         'genome_version': GENOME_VERSION_LOOKUP[genome_version],
         'sort': sort,
-        'sort_metadata': _get_sort_metadata(sort, sample_data),
+        'sort_metadata': _get_sort_metadata(sort, samples),
         'num_results': end_offset,
     }
     search_body.update(search)
@@ -30,12 +37,18 @@ def get_hail_variants(samples, search, user, previous_search_results, genome_ver
     })
     search_body.pop('skipped_samples', None)
 
+    search_body['sample_data'] = _get_sample_data(samples, search_body.get('inheritance_filter'))
+
     _parse_location_search(search_body)
 
     path = 'gene_counts' if gene_agg else 'search'
-    response = requests.post(f'http://hail-search:5000/{path}', json=search_body)
+    response = requests.post(_hail_backend_url(path), json=search_body, timeout=300)
     response.raise_for_status()
     response_json = response.json()
+
+    if gene_agg:
+        previous_search_results['gene_aggs'] = response_json
+        return response_json
 
     previous_search_results['total_results'] = response_json['total']
     previous_search_results['all_results'] = response_json['results']
@@ -43,7 +56,7 @@ def get_hail_variants(samples, search, user, previous_search_results, genome_ver
 
 
 def _get_sample_data(samples, inheritance_filter):
-    sample_data = samples.values(
+    sample_data = samples.order_by('id').values(
         'sample_id', 'dataset_type', 'sample_type',
         individual_guid=F('individual__guid'),
         family_guid=F('individual__family__guid'),
@@ -52,9 +65,10 @@ def _get_sample_data(samples, inheritance_filter):
         sex=F('individual__sex'),
     )
 
-    if (inheritance_filter or {}).get('affected'):
+    custom_affected = (inheritance_filter or {}).pop('affected', None)
+    if custom_affected:
         for s in sample_data:
-            s['affected'] = inheritance_filter['affected'].get(s['individual_guid']) or s['affected']
+            s['affected'] = custom_affected.get(s['individual_guid']) or s['affected']
 
     sample_data_by_data_type = defaultdict(list)
     for s in sample_data:
@@ -66,10 +80,10 @@ def _get_sample_data(samples, inheritance_filter):
     return sample_data_by_data_type
 
 
-def _get_sort_metadata(sort, sample_data):
+def _get_sort_metadata(sort, samples):
     sort_metadata = None
     if sort == 'in_omim':
-        sort_metadata = Omim.objects.filter(phenotype_mim_number__isnull=False).values_list('gene__gene_id', flat=True)
+        sort_metadata = list(Omim.objects.filter(phenotype_mim_number__isnull=False).values_list('gene__gene_id', flat=True))
     elif sort == 'constraint':
         sort_metadata = {
             agg['gene__gene_id']: agg['mis_z_rank'] + agg['pLI_rank'] for agg in
@@ -78,7 +92,7 @@ def _get_sort_metadata(sort, sample_data):
     elif sort == PRIORITIZED_GENE_SORT:
         sort_metadata = {
             agg['gene_id']: agg['min_rank'] for agg in PhenotypePrioritization.objects.filter(
-                individual__family__guid=next(sample_data.values())[0]['family_guid'], rank__lte=100,
+                individual__family_id=samples[0].individual.family_id, rank__lte=100,
             ).values('gene_id').annotate(min_rank=Min('rank'))
         }
     return sort_metadata
@@ -96,25 +110,23 @@ def _parse_location_search(search):
             {field: gene[f'{field}{search["genome_version"].title()}'] for field in ['chrom', 'start', 'end']}
             for gene in genes.values()
         ]
-        parsed_intervals = ['{chrom}:{start}-{end}'.format(**interval) for interval in intervals or []] + [
+        parsed_intervals = [_format_interval(**interval) for interval in intervals or []] + [
             '{chrom}:{start}-{end}'.format(**gene) for gene in gene_coords]
 
     exclude_locations = locus.get('excludeLocations')
-    has_location_search = bool(parsed_intervals) and not exclude_locations
-    if not has_location_search:
-        projects = set()
-        for samples in search['sample_data'].values():
-            projects.update({s['project_guid'] for s in samples})
-            if len(projects) > 1:
-                break
-        if len(projects) > 1:
-            from seqr.utils.search.utils import InvalidSearchException
-            raise InvalidSearchException('Location must be specified to search across multiple projects')
 
     search.update({
         'intervals': parsed_intervals,
         'exclude_intervals': exclude_locations,
-        'gene_ids': None if exclude_locations else list(genes.keys()),
+        'gene_ids': None if (exclude_locations or not genes) else list(genes.keys()),
         'variant_ids': parsed_locus.get('parsed_variant_ids'),
         'rs_ids': parsed_locus.get('rs_ids'),
     })
+
+
+def _format_interval(chrom=None, start=None, end=None, offset=None, **kwargs):
+    if offset:
+        offset_pos = int((end - start) * offset)
+        start = max(start - offset_pos, MIN_POS)
+        end = min(end + offset_pos, MAX_POS)
+    return f'{chrom}:{start}-{end}'