+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## dev
 
+## 6/2/23
+* Update Clinvar filtering and display
+
+* Add support for Gencode v39
+  * To add new data, run the `update_gencode_latest`
+
 ## 4/26/23
 * Add RnaSeqSpliceOutlier model (REQUIRES DB MIGRATION)
 * Add db index to improve Rna Sample Metadata performance (REQUIRES DB MIGRATION)